Struct rusoto_codepipeline::CreateCustomActionTypeInput

pub struct CreateCustomActionTypeInput {
    pub category: String,
    pub configuration_properties: Option<Vec<ActionConfigurationProperty>>,
    pub input_artifact_details: ArtifactDetails,
    pub output_artifact_details: ArtifactDetails,
    pub provider: String,
    pub settings: Option<ActionTypeSettings>,
    pub version: String,
}

Represents the input of a create custom action operation.

Fields
category: String

The category of the custom action, such as a build action or a test action.

Although Source and Approval are listed as valid values, they are not currently functional. These values are reserved for future use.

configuration_properties: Option<Vec<ActionConfigurationProperty>>

The configuration properties for the custom action.

You can refer to a name in the configuration properties of the custom action within the URL templates by following the format of {Config:name}, as long as the configuration property is both required and not secret. For more information, see Create a Custom Action for a Pipeline.

input_artifact_details: ArtifactDetails

The details of the input artifact for the action, such as its commit ID.

output_artifact_details: ArtifactDetails

The details of the output artifact of the action, such as its commit ID.

provider: String

The provider of the service used in the custom action, such as AWS CodeDeploy.

settings: Option<ActionTypeSettings>

Returns information about the settings for an action type.

version: String

The version identifier of the custom action.