Struct rusoto_codepipeline::ThirdPartyJobData

pub struct ThirdPartyJobData {
    pub action_configuration: Option<ActionConfiguration>,
    pub action_type_id: Option<ActionTypeId>,
    pub artifact_credentials: Option<AWSSessionCredentials>,
    pub continuation_token: Option<String>,
    pub encryption_key: Option<EncryptionKey>,
    pub input_artifacts: Option<Vec<Artifact>>,
    pub output_artifacts: Option<Vec<Artifact>>,
    pub pipeline_context: Option<PipelineContext>,
}

Represents information about the job data for a partner action.

Fields
action_configuration: Option<ActionConfiguration>

Represents information about an action configuration.

action_type_id: Option<ActionTypeId>

Represents information about an action type.

artifact_credentials: Option<AWSSessionCredentials>

Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the Amazon S3 bucket used to store artifact for the pipeline in AWS CodePipeline.

continuation_token: Option<String>

A system-generated token, such as a AWS CodeDeploy deployment ID, that a job requires in order to continue the job asynchronously.

encryption_key: Option<EncryptionKey>

The encryption key used to encrypt and decrypt data in the artifact store for the pipeline, such as an AWS Key Management Service (AWS KMS) key. This is optional and might not be present.

input_artifacts: Option<Vec<Artifact>>

The name of the artifact that will be worked upon by the action, if any. This name might be system-generated, such as "MyApp", or might be defined by the user when the action is created. The input artifact name must match the name of an output artifact generated by an action in an earlier action or stage of the pipeline.

output_artifacts: Option<Vec<Artifact>>

The name of the artifact that will be the result of the action, if any. This name might be system-generated, such as "MyBuiltApp", or might be defined by the user when the action is created.

pipeline_context: Option<PipelineContext>

Represents information about a pipeline to a job worker.