Trait rusoto_codepipeline::CodePipeline

pub trait CodePipeline {
    fn acknowledge_job(
        &self, 
        input: AcknowledgeJobInput
    ) -> RusotoFuture<AcknowledgeJobOutput, AcknowledgeJobError>;

    fn acknowledge_third_party_job(
        &self, 
        input: AcknowledgeThirdPartyJobInput
    ) -> RusotoFuture<AcknowledgeThirdPartyJobOutput, AcknowledgeThirdPartyJobError>;

    fn create_custom_action_type(
        &self, 
        input: CreateCustomActionTypeInput
    ) -> RusotoFuture<CreateCustomActionTypeOutput, CreateCustomActionTypeError>;

    fn create_pipeline(
        &self, 
        input: CreatePipelineInput
    ) -> RusotoFuture<CreatePipelineOutput, CreatePipelineError>;

    fn delete_custom_action_type(
        &self, 
        input: DeleteCustomActionTypeInput
    ) -> RusotoFuture<(), DeleteCustomActionTypeError>;

    fn delete_pipeline(
        &self, 
        input: DeletePipelineInput
    ) -> RusotoFuture<(), DeletePipelineError>;

    fn delete_webhook(
        &self, 
        input: DeleteWebhookInput
    ) -> RusotoFuture<DeleteWebhookOutput, DeleteWebhookError>;

    fn deregister_webhook_with_third_party(
        &self, 
        input: DeregisterWebhookWithThirdPartyInput
    ) -> RusotoFuture<DeregisterWebhookWithThirdPartyOutput, DeregisterWebhookWithThirdPartyError>;

    fn disable_stage_transition(
        &self, 
        input: DisableStageTransitionInput
    ) -> RusotoFuture<(), DisableStageTransitionError>;

    fn enable_stage_transition(
        &self, 
        input: EnableStageTransitionInput
    ) -> RusotoFuture<(), EnableStageTransitionError>;

    fn get_job_details(
        &self, 
        input: GetJobDetailsInput
    ) -> RusotoFuture<GetJobDetailsOutput, GetJobDetailsError>;

    fn get_pipeline(
        &self, 
        input: GetPipelineInput
    ) -> RusotoFuture<GetPipelineOutput, GetPipelineError>;

    fn get_pipeline_execution(
        &self, 
        input: GetPipelineExecutionInput
    ) -> RusotoFuture<GetPipelineExecutionOutput, GetPipelineExecutionError>;

    fn get_pipeline_state(
        &self, 
        input: GetPipelineStateInput
    ) -> RusotoFuture<GetPipelineStateOutput, GetPipelineStateError>;

    fn get_third_party_job_details(
        &self, 
        input: GetThirdPartyJobDetailsInput
    ) -> RusotoFuture<GetThirdPartyJobDetailsOutput, GetThirdPartyJobDetailsError>;

    fn list_action_executions(
        &self, 
        input: ListActionExecutionsInput
    ) -> RusotoFuture<ListActionExecutionsOutput, ListActionExecutionsError>;

    fn list_action_types(
        &self, 
        input: ListActionTypesInput
    ) -> RusotoFuture<ListActionTypesOutput, ListActionTypesError>;

    fn list_pipeline_executions(
        &self, 
        input: ListPipelineExecutionsInput
    ) -> RusotoFuture<ListPipelineExecutionsOutput, ListPipelineExecutionsError>;

    fn list_pipelines(
        &self, 
        input: ListPipelinesInput
    ) -> RusotoFuture<ListPipelinesOutput, ListPipelinesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceInput
    ) -> RusotoFuture<ListTagsForResourceOutput, ListTagsForResourceError>;

    fn list_webhooks(
        &self, 
        input: ListWebhooksInput
    ) -> RusotoFuture<ListWebhooksOutput, ListWebhooksError>;

    fn poll_for_jobs(
        &self, 
        input: PollForJobsInput
    ) -> RusotoFuture<PollForJobsOutput, PollForJobsError>;

    fn poll_for_third_party_jobs(
        &self, 
        input: PollForThirdPartyJobsInput
    ) -> RusotoFuture<PollForThirdPartyJobsOutput, PollForThirdPartyJobsError>;

    fn put_action_revision(
        &self, 
        input: PutActionRevisionInput
    ) -> RusotoFuture<PutActionRevisionOutput, PutActionRevisionError>;

    fn put_approval_result(
        &self, 
        input: PutApprovalResultInput
    ) -> RusotoFuture<PutApprovalResultOutput, PutApprovalResultError>;

    fn put_job_failure_result(
        &self, 
        input: PutJobFailureResultInput
    ) -> RusotoFuture<(), PutJobFailureResultError>;

    fn put_job_success_result(
        &self, 
        input: PutJobSuccessResultInput
    ) -> RusotoFuture<(), PutJobSuccessResultError>;

    fn put_third_party_job_failure_result(
        &self, 
        input: PutThirdPartyJobFailureResultInput
    ) -> RusotoFuture<(), PutThirdPartyJobFailureResultError>;

    fn put_third_party_job_success_result(
        &self, 
        input: PutThirdPartyJobSuccessResultInput
    ) -> RusotoFuture<(), PutThirdPartyJobSuccessResultError>;

    fn put_webhook(
        &self, 
        input: PutWebhookInput
    ) -> RusotoFuture<PutWebhookOutput, PutWebhookError>;

    fn register_webhook_with_third_party(
        &self, 
        input: RegisterWebhookWithThirdPartyInput
    ) -> RusotoFuture<RegisterWebhookWithThirdPartyOutput, RegisterWebhookWithThirdPartyError>;

    fn retry_stage_execution(
        &self, 
        input: RetryStageExecutionInput
    ) -> RusotoFuture<RetryStageExecutionOutput, RetryStageExecutionError>;

    fn start_pipeline_execution(
        &self, 
        input: StartPipelineExecutionInput
    ) -> RusotoFuture<StartPipelineExecutionOutput, StartPipelineExecutionError>;

    fn tag_resource(
        &self, 
        input: TagResourceInput
    ) -> RusotoFuture<TagResourceOutput, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceInput
    ) -> RusotoFuture<UntagResourceOutput, UntagResourceError>;

    fn update_pipeline(
        &self, 
        input: UpdatePipelineInput
    ) -> RusotoFuture<UpdatePipelineOutput, UpdatePipelineError>;
}

Trait representing the capabilities of the CodePipeline API. CodePipeline clients implement this trait.

Required methods

fn acknowledge_job(
&self,
input: AcknowledgeJobInput
) -> RusotoFuture<AcknowledgeJobOutput, AcknowledgeJobError>

Returns information about a specified job and whether that job has been received by the job worker. Only used for custom actions.

fn acknowledge_third_party_job(
&self,
input: AcknowledgeThirdPartyJobInput
) -> RusotoFuture<AcknowledgeThirdPartyJobOutput, AcknowledgeThirdPartyJobError>

Confirms a job worker has received the specified job. Only used for partner actions.

fn create_custom_action_type(
&self,
input: CreateCustomActionTypeInput
) -> RusotoFuture<CreateCustomActionTypeOutput, CreateCustomActionTypeError>

Creates a new custom action that can be used in all pipelines associated with the AWS account. Only used for custom actions.

fn create_pipeline(
&self,
input: CreatePipelineInput
) -> RusotoFuture<CreatePipelineOutput, CreatePipelineError>

Creates a pipeline.

fn delete_custom_action_type(
&self,
input: DeleteCustomActionTypeInput
) -> RusotoFuture<(), DeleteCustomActionTypeError>

Marks a custom action as deleted. PollForJobs for the custom action will fail after the action is marked for deletion. Only used for custom actions.

To re-create a custom action after it has been deleted you must use a string in the version field that has never been used before. This string can be an incremented version number, for example. To restore a deleted custom action, use a JSON file that is identical to the deleted action, including the original string in the version field.

fn delete_pipeline(
&self,
input: DeletePipelineInput
) -> RusotoFuture<(), DeletePipelineError>

Deletes the specified pipeline.

fn delete_webhook(
&self,
input: DeleteWebhookInput
) -> RusotoFuture<DeleteWebhookOutput, DeleteWebhookError>

Deletes a previously created webhook by name. Deleting the webhook stops AWS CodePipeline from starting a pipeline every time an external event occurs. The API will return successfully when trying to delete a webhook that is already deleted. If a deleted webhook is re-created by calling PutWebhook with the same name, it will have a different URL.

fn deregister_webhook_with_third_party(
&self,
input: DeregisterWebhookWithThirdPartyInput
) -> RusotoFuture<DeregisterWebhookWithThirdPartyOutput, DeregisterWebhookWithThirdPartyError>

Removes the connection between the webhook that was created by CodePipeline and the external tool with events to be detected. Currently only supported for webhooks that target an action type of GitHub.

fn disable_stage_transition(
&self,
input: DisableStageTransitionInput
) -> RusotoFuture<(), DisableStageTransitionError>

Prevents artifacts in a pipeline from transitioning to the next stage in the pipeline.

fn enable_stage_transition(
&self,
input: EnableStageTransitionInput
) -> RusotoFuture<(), EnableStageTransitionError>

Enables artifacts in a pipeline to transition to a stage in a pipeline.

fn get_job_details(
&self,
input: GetJobDetailsInput
) -> RusotoFuture<GetJobDetailsOutput, GetJobDetailsError>

Returns information about a job. Only used for custom actions.

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

fn get_pipeline(
&self,
input: GetPipelineInput
) -> RusotoFuture<GetPipelineOutput, GetPipelineError>

Returns the metadata, structure, stages, and actions of a pipeline. Can be used to return the entire structure of a pipeline in JSON format, which can then be modified and used to update the pipeline structure with UpdatePipeline.

fn get_pipeline_execution(
&self,
input: GetPipelineExecutionInput
) -> RusotoFuture<GetPipelineExecutionOutput, GetPipelineExecutionError>

Returns information about an execution of a pipeline, including details about artifacts, the pipeline execution ID, and the name, version, and status of the pipeline.

fn get_pipeline_state(
&self,
input: GetPipelineStateInput
) -> RusotoFuture<GetPipelineStateOutput, GetPipelineStateError>

Returns information about the state of a pipeline, including the stages and actions.

Values returned in the revisionId and revisionUrl fields indicate the source revision information, such as the commit ID, for the current state.

fn get_third_party_job_details(
&self,
input: GetThirdPartyJobDetailsInput
) -> RusotoFuture<GetThirdPartyJobDetailsOutput, GetThirdPartyJobDetailsError>

Requests the details of a job for a third party action. Only used for partner actions.

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

fn list_action_executions(
&self,
input: ListActionExecutionsInput
) -> RusotoFuture<ListActionExecutionsOutput, ListActionExecutionsError>

Lists the action executions that have occurred in a pipeline.

fn list_action_types(
&self,
input: ListActionTypesInput
) -> RusotoFuture<ListActionTypesOutput, ListActionTypesError>

Gets a summary of all AWS CodePipeline action types associated with your account.

fn list_pipeline_executions(
&self,
input: ListPipelineExecutionsInput
) -> RusotoFuture<ListPipelineExecutionsOutput, ListPipelineExecutionsError>

Gets a summary of the most recent executions for a pipeline.

fn list_pipelines(
&self,
input: ListPipelinesInput
) -> RusotoFuture<ListPipelinesOutput, ListPipelinesError>

Gets a summary of all of the pipelines associated with your account.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceInput
) -> RusotoFuture<ListTagsForResourceOutput, ListTagsForResourceError>

Gets the set of key/value pairs (metadata) that are used to manage the resource.

fn list_webhooks(
&self,
input: ListWebhooksInput
) -> RusotoFuture<ListWebhooksOutput, ListWebhooksError>

Gets a listing of all the webhooks in this region for this account. The output lists all webhooks and includes the webhook URL and ARN, as well the configuration for each webhook.

fn poll_for_jobs(
&self,
input: PollForJobsInput
) -> RusotoFuture<PollForJobsOutput, PollForJobsError>

Returns information about any jobs for AWS CodePipeline to act upon. PollForJobs is only valid for action types with "Custom" in the owner field. If the action type contains "AWS" or "ThirdParty" in the owner field, the PollForJobs action returns an error.

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

fn poll_for_third_party_jobs(
&self,
input: PollForThirdPartyJobsInput
) -> RusotoFuture<PollForThirdPartyJobsOutput, PollForThirdPartyJobsError>

Determines whether there are any third party jobs for a job worker to act on. Only used for partner actions.

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts.

fn put_action_revision(
&self,
input: PutActionRevisionInput
) -> RusotoFuture<PutActionRevisionOutput, PutActionRevisionError>

Provides information to AWS CodePipeline about new revisions to a source.

fn put_approval_result(
&self,
input: PutApprovalResultInput
) -> RusotoFuture<PutApprovalResultOutput, PutApprovalResultError>

Provides the response to a manual approval request to AWS CodePipeline. Valid responses include Approved and Rejected.

fn put_job_failure_result(
&self,
input: PutJobFailureResultInput
) -> RusotoFuture<(), PutJobFailureResultError>

Represents the failure of a job as returned to the pipeline by a job worker. Only used for custom actions.

fn put_job_success_result(
&self,
input: PutJobSuccessResultInput
) -> RusotoFuture<(), PutJobSuccessResultError>

Represents the success of a job as returned to the pipeline by a job worker. Only used for custom actions.

fn put_third_party_job_failure_result(
&self,
input: PutThirdPartyJobFailureResultInput
) -> RusotoFuture<(), PutThirdPartyJobFailureResultError>

Represents the failure of a third party job as returned to the pipeline by a job worker. Only used for partner actions.

fn put_third_party_job_success_result(
&self,
input: PutThirdPartyJobSuccessResultInput
) -> RusotoFuture<(), PutThirdPartyJobSuccessResultError>

Represents the success of a third party job as returned to the pipeline by a job worker. Only used for partner actions.

fn put_webhook(
&self,
input: PutWebhookInput
) -> RusotoFuture<PutWebhookOutput, PutWebhookError>

Defines a webhook and returns a unique webhook URL generated by CodePipeline. This URL can be supplied to third party source hosting providers to call every time there's a code change. When CodePipeline receives a POST request on this URL, the pipeline defined in the webhook is started as long as the POST request satisfied the authentication and filtering requirements supplied when defining the webhook. RegisterWebhookWithThirdParty and DeregisterWebhookWithThirdParty APIs can be used to automatically configure supported third parties to call the generated webhook URL.

fn register_webhook_with_third_party(
&self,
input: RegisterWebhookWithThirdPartyInput
) -> RusotoFuture<RegisterWebhookWithThirdPartyOutput, RegisterWebhookWithThirdPartyError>

Configures a connection between the webhook that was created and the external tool with events to be detected.

fn retry_stage_execution(
&self,
input: RetryStageExecutionInput
) -> RusotoFuture<RetryStageExecutionOutput, RetryStageExecutionError>

Resumes the pipeline execution by retrying the last failed actions in a stage.

fn start_pipeline_execution(
&self,
input: StartPipelineExecutionInput
) -> RusotoFuture<StartPipelineExecutionOutput, StartPipelineExecutionError>

Starts the specified pipeline. Specifically, it begins processing the latest commit to the source location specified as part of the pipeline.

fn tag_resource(
&self,
input: TagResourceInput
) -> RusotoFuture<TagResourceOutput, TagResourceError>

Adds to or modifies the tags of the given resource. Tags are metadata that can be used to manage a resource.

fn untag_resource(
&self,
input: UntagResourceInput
) -> RusotoFuture<UntagResourceOutput, UntagResourceError>

Removes tags from an AWS resource.

fn update_pipeline(
&self,
input: UpdatePipelineInput
) -> RusotoFuture<UpdatePipelineOutput, UpdatePipelineError>

Updates a specified pipeline with edits or changes to its structure. Use a JSON file with the pipeline structure in conjunction with UpdatePipeline to provide the full structure of the pipeline. Updating the pipeline increases the version number of the pipeline by 1.

Loading content...