Crate rusoto_codepipeline

AWS CodePipeline

If you're using the service, you're probably looking for CodePipelineClient and CodePipeline.

Structs

AWSSessionCredentials
Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the Amazon S3 bucket used to store artifact for the pipeline in AWS CodePipeline.

AcknowledgeJobInput
Represents the input of an acknowledge job action.

AcknowledgeJobOutput
Represents the output of an acknowledge job action.

AcknowledgeThirdPartyJobInput
Represents the input of an acknowledge third party job action.

AcknowledgeThirdPartyJobOutput
Represents the output of an acknowledge third party job action.

ActionConfiguration
Represents information about an action configuration.

ActionConfigurationProperty
Represents information about an action configuration property.

ActionContext
Represents the context of an action within the stage of a pipeline to a job worker.

ActionDeclaration
Represents information about an action declaration.

ActionExecution
Represents information about the run of an action.

ActionRevision
Represents information about the version (or revision) of an action.

ActionState
Represents information about the state of an action.

ActionType
Returns information about the details of an action type.

ActionTypeId
Represents information about an action type.

ActionTypeSettings
Returns information about the settings for an action type.

ApprovalResult
Represents information about the result of an approval request.

Artifact
Represents information about an artifact that will be worked upon by actions in the pipeline.

ArtifactDetails
Returns information about the details of an artifact.

ArtifactLocation
Represents information about the location of an artifact.

ArtifactRevision
Represents revision details of an artifact.

ArtifactStore
The Amazon S3 location where artifacts are stored for the pipeline. If this Amazon S3 bucket is created manually, it must meet the requirements for AWS CodePipeline. For more information, see the Concepts.

BlockerDeclaration
Reserved for future use.

CodePipelineClient
A client for the CodePipeline API.

CreateCustomActionTypeInput
Represents the input of a create custom action operation.

CreateCustomActionTypeOutput
Represents the output of a create custom action operation.

CreatePipelineInput
Represents the input of a create pipeline action.

CreatePipelineOutput
Represents the output of a create pipeline action.

CurrentRevision
Represents information about a current revision.

DeleteCustomActionTypeInput
Represents the input of a delete custom action operation. The custom action will be marked as deleted.

DeletePipelineInput
Represents the input of a delete pipeline action.

DisableStageTransitionInput
Represents the input of a disable stage transition input action.

EnableStageTransitionInput
Represents the input of an enable stage transition action.

EncryptionKey
Represents information about the key used to encrypt data in the artifact store, such as an AWS Key Management Service (AWS KMS) key.

ErrorDetails
Represents information about an error in AWS CodePipeline.

ExecutionDetails
The details of the actions taken and results produced on an artifact as it passes through stages in the pipeline.

FailureDetails
Represents information about failure details.

GetJobDetailsInput
Represents the input of a get job details action.

GetJobDetailsOutput
Represents the output of a get job details action.

GetPipelineExecutionInput
Represents the input of a get pipeline execution action.

GetPipelineExecutionOutput
Represents the output of a get pipeline execution action.

GetPipelineInput
Represents the input of a get pipeline action.

GetPipelineOutput
Represents the output of a get pipeline action.

GetPipelineStateInput
Represents the input of a get pipeline state action.

GetPipelineStateOutput
Represents the output of a get pipeline state action.

GetThirdPartyJobDetailsInput
Represents the input of a get third party job details action.

GetThirdPartyJobDetailsOutput
Represents the output of a get third party job details action.

InputArtifact
Represents information about an artifact to be worked on, such as a test or build artifact.

Job
Represents information about a job.

JobData
Represents additional information about a job required for a job worker to complete the job.

JobDetails
Represents information about the details of a job.

ListActionTypesInput
Represents the input of a list action types action.

ListActionTypesOutput
Represents the output of a list action types action.

ListPipelinesInput
Represents the input of a list pipelines action.

ListPipelinesOutput
Represents the output of a list pipelines action.

OutputArtifact
Represents information about the output of an action.

PipelineContext
Represents information about a pipeline to a job worker.

PipelineDeclaration
Represents the structure of actions and stages to be performed in the pipeline.

PipelineExecution
Represents information about an execution of a pipeline.

PipelineSummary
Returns a summary of a pipeline.

PollForJobsInput
Represents the input of a poll for jobs action.

PollForJobsOutput
Represents the output of a poll for jobs action.

PollForThirdPartyJobsInput
Represents the input of a poll for third party jobs action.

PollForThirdPartyJobsOutput
Represents the output of a poll for third party jobs action.

PutActionRevisionInput
Represents the input of a put action revision action.

PutActionRevisionOutput
Represents the output of a put action revision action.

PutApprovalResultInput
Represents the input of a put approval result action.

PutApprovalResultOutput
Represents the output of a put approval result action.

PutJobFailureResultInput
Represents the input of a put job failure result action.

PutJobSuccessResultInput
Represents the input of a put job success result action.

PutThirdPartyJobFailureResultInput
Represents the input of a third party job failure result action.

PutThirdPartyJobSuccessResultInput
Represents the input of a put third party job success result action.

RetryStageExecutionInput
Represents the input of a retry stage execution action.

RetryStageExecutionOutput
Represents the output of a retry stage execution action.

S3ArtifactLocation
The location of the Amazon S3 bucket that contains a revision.

StageContext
Represents information about a stage to a job worker.

StageDeclaration
Represents information about a stage and its definition.

StageExecution
Represents information about the run of a stage.

StageState
Represents information about the state of the stage.

StartPipelineExecutionInput
Represents the input of a start pipeline execution action.

StartPipelineExecutionOutput
Represents the output of a start pipeline execution action.

ThirdPartyJob
A response to a PollForThirdPartyJobs request returned by AWS CodePipeline when there is a job to be worked upon by a partner action.

ThirdPartyJobData
Represents information about the job data for a partner action.

ThirdPartyJobDetails
The details of a job sent in response to a GetThirdPartyJobDetails request.

TransitionState
Represents information about the state of transitions between one stage and another stage.

UpdatePipelineInput
Represents the input of an update pipeline action.

UpdatePipelineOutput
Represents the output of an update pipeline action.

Enums

AcknowledgeJobError
Errors returned by AcknowledgeJob

AcknowledgeThirdPartyJobError
Errors returned by AcknowledgeThirdPartyJob

CreateCustomActionTypeError
Errors returned by CreateCustomActionType

CreatePipelineError
Errors returned by CreatePipeline

DeleteCustomActionTypeError
Errors returned by DeleteCustomActionType

DeletePipelineError
Errors returned by DeletePipeline

DisableStageTransitionError
Errors returned by DisableStageTransition

EnableStageTransitionError
Errors returned by EnableStageTransition

GetJobDetailsError
Errors returned by GetJobDetails

GetPipelineError
Errors returned by GetPipeline

GetPipelineExecutionError
Errors returned by GetPipelineExecution

GetPipelineStateError
Errors returned by GetPipelineState

GetThirdPartyJobDetailsError
Errors returned by GetThirdPartyJobDetails

ListActionTypesError
Errors returned by ListActionTypes

ListPipelinesError
Errors returned by ListPipelines

PollForJobsError
Errors returned by PollForJobs

PollForThirdPartyJobsError
Errors returned by PollForThirdPartyJobs

PutActionRevisionError
Errors returned by PutActionRevision

PutApprovalResultError
Errors returned by PutApprovalResult

PutJobFailureResultError
Errors returned by PutJobFailureResult

PutJobSuccessResultError
Errors returned by PutJobSuccessResult

PutThirdPartyJobFailureResultError
Errors returned by PutThirdPartyJobFailureResult

PutThirdPartyJobSuccessResultError
Errors returned by PutThirdPartyJobSuccessResult

RetryStageExecutionError
Errors returned by RetryStageExecution

StartPipelineExecutionError
Errors returned by StartPipelineExecution

UpdatePipelineError
Errors returned by UpdatePipeline

Traits

CodePipeline
Trait representing the capabilities of the CodePipeline API. CodePipeline clients implement this trait.

Type Definitions

AccessKeyId

AccountId

ActionCategory

ActionConfigurationKey

ActionConfigurationMap

ActionConfigurationPropertyList

ActionConfigurationPropertyType

ActionConfigurationQueryableValue

ActionConfigurationValue

ActionExecutionStatus

ActionExecutionToken

ActionName

ActionOwner

ActionProvider

ActionRunOrder

ActionStateList

ActionTypeList

ApprovalStatus

ApprovalSummary

ApprovalToken

ArtifactList

ArtifactLocationType

ArtifactName

ArtifactRevisionList

ArtifactStoreLocation

ArtifactStoreType

BlockerName

BlockerType

Boolean

ClientId

ClientToken

Code

ContinuationToken

Description

DisabledReason

Enabled

EncryptionKeyId

EncryptionKeyType

ExecutionId

ExecutionSummary

FailureType

InputArtifactList

JobId

JobList

JobStatus

LastChangedAt

LastChangedBy

LastUpdatedBy

MaxBatchSize

MaximumArtifactCount

Message

MinimumArtifactCount

NextToken

Nonce

OutputArtifactList

Percentage

PipelineExecutionId

PipelineExecutionStatus

PipelineList

PipelineName

PipelineStageDeclarationList

PipelineVersion

QueryParamMap

Revision

RevisionChangeIdentifier

RevisionSummary

RoleArn

S3BucketName

S3ObjectKey

SecretAccessKey

SessionToken

StageActionDeclarationList

StageBlockerDeclarationList

StageExecutionStatus

StageName

StageRetryMode

StageStateList

StageTransitionType

ThirdPartyJobId

ThirdPartyJobList

Time

Timestamp

Url

UrlTemplate

Version