Crate rusoto_codepipeline

AWS CodePipeline
Overview

This is the AWS CodePipeline API Reference. This guide provides descriptions of the actions and data types for AWS CodePipeline. Some functionality for your pipeline is only configurable through the API. For additional information, see the AWS CodePipeline User Guide.

You can use the AWS CodePipeline API to work with pipelines, stages, actions, and transitions, as described below.

Pipelines are models of automated release processes. Each pipeline is uniquely named, and consists of stages, actions, and transitions.

You can work with pipelines by calling:

CreatePipeline, which creates a uniquely-named pipeline.

DeletePipeline, which deletes the specified pipeline.

GetPipeline, which returns information about the pipeline structure and pipeline metadata, including the pipeline Amazon Resource Name (ARN).

GetPipelineExecution, which returns information about a specific execution of a pipeline.

GetPipelineState, which returns information about the current state of the stages and actions of a pipeline.

ListPipelines, which gets a summary of all of the pipelines associated with your account.

ListPipelineExecutions, which gets a summary of the most recent executions for a pipeline.

StartPipelineExecution, which runs the the most recent revision of an artifact through the pipeline.

UpdatePipeline, which updates a pipeline with edits or changes to the structure of the pipeline.

Pipelines include stages. Each stage contains one or more actions that must complete before the next stage begins. A stage will result in success or failure. If a stage fails, then the pipeline stops at that stage and will remain stopped until either a new version of an artifact appears in the source location, or a user takes action to re-run the most recent artifact through the pipeline. You can call GetPipelineState, which displays the status of a pipeline, including the status of stages in the pipeline, or GetPipeline, which returns the entire structure of the pipeline, including the stages of that pipeline. For more information about the structure of stages and actions, also refer to the AWS CodePipeline Pipeline Structure Reference.

Pipeline stages include actions, which are categorized into categories such as source or build actions performed within a stage of a pipeline. For example, you can use a source action to import artifacts into a pipeline from a source such as Amazon S3. Like stages, you do not work with actions directly in most cases, but you do define and interact with actions when working with pipeline operations such as CreatePipeline and GetPipelineState. Valid action categories are:

Source

Build

Test

Deploy

Approval

Invoke

Pipelines also include transitions, which allow the transition of artifacts from one stage to the next in a pipeline after the actions in one stage complete.

You can work with transitions by calling:

DisableStageTransition, which prevents artifacts from transitioning to the next stage in a pipeline.

EnableStageTransition, which enables transition of artifacts between stages in a pipeline.

Using the API to integrate with AWS CodePipeline

For third-party integrators or developers who want to create their own integrations with AWS CodePipeline, the expected sequence varies from the standard API user. In order to integrate with AWS CodePipeline, developers will need to work with the following items:

Jobs, which are instances of an action. For example, a job for a source action might import a revision of an artifact from a source.

You can work with jobs by calling:

AcknowledgeJob, which confirms whether a job worker has received the specified job,

GetJobDetails, which returns the details of a job,

PollForJobs, which determines whether there are any jobs to act upon,

PutJobFailureResult, which provides details of a job failure, and

PutJobSuccessResult, which provides details of a job success.

Third party jobs, which are instances of an action created by a partner action and integrated into AWS CodePipeline. Partner actions are created by members of the AWS Partner Network.

You can work with third party jobs by calling:

AcknowledgeThirdPartyJob, which confirms whether a job worker has received the specified job,

GetThirdPartyJobDetails, which requests the details of a job for a partner action,

PollForThirdPartyJobs, which determines whether there are any jobs to act upon,

PutThirdPartyJobFailureResult, which provides details of a job failure, and

PutThirdPartyJobSuccessResult, which provides details of a job success.

If you're using the service, you're probably looking for CodePipelineClient and CodePipeline.

Structs

AWSSessionCredentials
Represents an AWS session credentials object. These credentials are temporary credentials that are issued by AWS Secure Token Service (STS). They can be used to access input and output artifacts in the Amazon S3 bucket used to store artifact for the pipeline in AWS CodePipeline.

AcknowledgeJobInput
Represents the input of an AcknowledgeJob action.

AcknowledgeJobOutput
Represents the output of an AcknowledgeJob action.

AcknowledgeThirdPartyJobInput
Represents the input of an AcknowledgeThirdPartyJob action.

AcknowledgeThirdPartyJobOutput
Represents the output of an AcknowledgeThirdPartyJob action.

ActionConfiguration
Represents information about an action configuration.

ActionConfigurationProperty
Represents information about an action configuration property.

ActionContext
Represents the context of an action within the stage of a pipeline to a job worker.

ActionDeclaration
Represents information about an action declaration.

ActionExecution
Represents information about the run of an action.

ActionRevision
Represents information about the version (or revision) of an action.

ActionState
Represents information about the state of an action.

ActionType
Returns information about the details of an action type.

ActionTypeId
Represents information about an action type.

ActionTypeSettings
Returns information about the settings for an action type.

ApprovalResult
Represents information about the result of an approval request.

Artifact
Represents information about an artifact that will be worked upon by actions in the pipeline.

ArtifactDetails
Returns information about the details of an artifact.

ArtifactLocation
Represents information about the location of an artifact.

ArtifactRevision
Represents revision details of an artifact.

ArtifactStore
The Amazon S3 bucket where artifacts are stored for the pipeline.

BlockerDeclaration
Reserved for future use.

CodePipelineClient
A client for the CodePipeline API.

CreateCustomActionTypeInput
Represents the input of a CreateCustomActionType operation.

CreateCustomActionTypeOutput
Represents the output of a CreateCustomActionType operation.

CreatePipelineInput
Represents the input of a CreatePipeline action.

CreatePipelineOutput
Represents the output of a CreatePipeline action.

CurrentRevision
Represents information about a current revision.

DeleteCustomActionTypeInput
Represents the input of a DeleteCustomActionType operation. The custom action will be marked as deleted.

DeletePipelineInput
Represents the input of a DeletePipeline action.

DeleteWebhookInput

DeleteWebhookOutput

DeregisterWebhookWithThirdPartyInput

DeregisterWebhookWithThirdPartyOutput

DisableStageTransitionInput
Represents the input of a DisableStageTransition action.

EnableStageTransitionInput
Represents the input of an EnableStageTransition action.

EncryptionKey
Represents information about the key used to encrypt data in the artifact store, such as an AWS Key Management Service (AWS KMS) key.

ErrorDetails
Represents information about an error in AWS CodePipeline.

ExecutionDetails
The details of the actions taken and results produced on an artifact as it passes through stages in the pipeline.

FailureDetails
Represents information about failure details.

GetJobDetailsInput
Represents the input of a GetJobDetails action.

GetJobDetailsOutput
Represents the output of a GetJobDetails action.

GetPipelineExecutionInput
Represents the input of a GetPipelineExecution action.

GetPipelineExecutionOutput
Represents the output of a GetPipelineExecution action.

GetPipelineInput
Represents the input of a GetPipeline action.

GetPipelineOutput
Represents the output of a GetPipeline action.

GetPipelineStateInput
Represents the input of a GetPipelineState action.

GetPipelineStateOutput
Represents the output of a GetPipelineState action.

GetThirdPartyJobDetailsInput
Represents the input of a GetThirdPartyJobDetails action.

GetThirdPartyJobDetailsOutput
Represents the output of a GetThirdPartyJobDetails action.

InputArtifact
Represents information about an artifact to be worked on, such as a test or build artifact.

Job
Represents information about a job.

JobData
Represents additional information about a job required for a job worker to complete the job.

JobDetails
Represents information about the details of a job.

ListActionTypesInput
Represents the input of a ListActionTypes action.

ListActionTypesOutput
Represents the output of a ListActionTypes action.

ListPipelineExecutionsInput
Represents the input of a ListPipelineExecutions action.

ListPipelineExecutionsOutput
Represents the output of a ListPipelineExecutions action.

ListPipelinesInput
Represents the input of a ListPipelines action.

ListPipelinesOutput
Represents the output of a ListPipelines action.

ListWebhookItem
The detail returned for each webhook after listing webhooks, such as the webhook URL, the webhook name, and the webhook ARN.

ListWebhooksInput

ListWebhooksOutput

OutputArtifact
Represents information about the output of an action.

PipelineContext
Represents information about a pipeline to a job worker.

PipelineDeclaration
Represents the structure of actions and stages to be performed in the pipeline.

PipelineExecution
Represents information about an execution of a pipeline.

PipelineExecutionSummary
Summary information about a pipeline execution.

PipelineMetadata
Information about a pipeline.

PipelineSummary
Returns a summary of a pipeline.

PollForJobsInput
Represents the input of a PollForJobs action.

PollForJobsOutput
Represents the output of a PollForJobs action.

PollForThirdPartyJobsInput
Represents the input of a PollForThirdPartyJobs action.

PollForThirdPartyJobsOutput
Represents the output of a PollForThirdPartyJobs action.

PutActionRevisionInput
Represents the input of a PutActionRevision action.

PutActionRevisionOutput
Represents the output of a PutActionRevision action.

PutApprovalResultInput
Represents the input of a PutApprovalResult action.

PutApprovalResultOutput
Represents the output of a PutApprovalResult action.

PutJobFailureResultInput
Represents the input of a PutJobFailureResult action.

PutJobSuccessResultInput
Represents the input of a PutJobSuccessResult action.

PutThirdPartyJobFailureResultInput
Represents the input of a PutThirdPartyJobFailureResult action.

PutThirdPartyJobSuccessResultInput
Represents the input of a PutThirdPartyJobSuccessResult action.

PutWebhookInput

PutWebhookOutput

RegisterWebhookWithThirdPartyInput

RegisterWebhookWithThirdPartyOutput

RetryStageExecutionInput
Represents the input of a RetryStageExecution action.

RetryStageExecutionOutput
Represents the output of a RetryStageExecution action.

S3ArtifactLocation
The location of the Amazon S3 bucket that contains a revision.

SourceRevision

StageContext
Represents information about a stage to a job worker.

StageDeclaration
Represents information about a stage and its definition.

StageExecution
Represents information about the run of a stage.

StageState
Represents information about the state of the stage.

StartPipelineExecutionInput
Represents the input of a StartPipelineExecution action.

StartPipelineExecutionOutput
Represents the output of a StartPipelineExecution action.

ThirdPartyJob
A response to a PollForThirdPartyJobs request returned by AWS CodePipeline when there is a job to be worked upon by a partner action.

ThirdPartyJobData
Represents information about the job data for a partner action.

ThirdPartyJobDetails
The details of a job sent in response to a GetThirdPartyJobDetails request.

TransitionState
Represents information about the state of transitions between one stage and another stage.

UpdatePipelineInput
Represents the input of an UpdatePipeline action.

UpdatePipelineOutput
Represents the output of an UpdatePipeline action.

WebhookAuthConfiguration

WebhookDefinition
Represents information about a webhook and its definition.

WebhookFilterRule
The event criteria that specify when a webhook notification is sent to your URL.

Enums

AcknowledgeJobError
Errors returned by AcknowledgeJob

AcknowledgeThirdPartyJobError
Errors returned by AcknowledgeThirdPartyJob

CreateCustomActionTypeError
Errors returned by CreateCustomActionType

CreatePipelineError
Errors returned by CreatePipeline

DeleteCustomActionTypeError
Errors returned by DeleteCustomActionType

DeletePipelineError
Errors returned by DeletePipeline

DeleteWebhookError
Errors returned by DeleteWebhook

DeregisterWebhookWithThirdPartyError
Errors returned by DeregisterWebhookWithThirdParty

DisableStageTransitionError
Errors returned by DisableStageTransition

EnableStageTransitionError
Errors returned by EnableStageTransition

GetJobDetailsError
Errors returned by GetJobDetails

GetPipelineError
Errors returned by GetPipeline

GetPipelineExecutionError
Errors returned by GetPipelineExecution

GetPipelineStateError
Errors returned by GetPipelineState

GetThirdPartyJobDetailsError
Errors returned by GetThirdPartyJobDetails

ListActionTypesError
Errors returned by ListActionTypes

ListPipelineExecutionsError
Errors returned by ListPipelineExecutions

ListPipelinesError
Errors returned by ListPipelines

ListWebhooksError
Errors returned by ListWebhooks

PollForJobsError
Errors returned by PollForJobs

PollForThirdPartyJobsError
Errors returned by PollForThirdPartyJobs

PutActionRevisionError
Errors returned by PutActionRevision

PutApprovalResultError
Errors returned by PutApprovalResult

PutJobFailureResultError
Errors returned by PutJobFailureResult

PutJobSuccessResultError
Errors returned by PutJobSuccessResult

PutThirdPartyJobFailureResultError
Errors returned by PutThirdPartyJobFailureResult

PutThirdPartyJobSuccessResultError
Errors returned by PutThirdPartyJobSuccessResult

PutWebhookError
Errors returned by PutWebhook

RegisterWebhookWithThirdPartyError
Errors returned by RegisterWebhookWithThirdParty

RetryStageExecutionError
Errors returned by RetryStageExecution

StartPipelineExecutionError
Errors returned by StartPipelineExecution

UpdatePipelineError
Errors returned by UpdatePipeline

Traits

CodePipeline
Trait representing the capabilities of the CodePipeline API. CodePipeline clients implement this trait.